Refined Render Graph System along with minor changes to fully support Blur Outline Render Graph.

+++ shaders/HLSL/BlurOutline_PS.hlsl
@@ -13,7 +13,7 @@ cbuffer Control : register(b1)
 
 float4 main(float2 uv : Texcoord) : SV_Target
 {
-    uint width, height;
+    float width, height;
     tex.GetDimensions(width, height);
     float dx, dy;
     if (horizontal)
